[DESKTOP VERSION] :: Fixed Design Issues
Fixed all the design issues with the website and aligned the front-end code to replicate the desgin patterns as per the PSD.

Rewrote the CSS from scratch to enfore DRY. Shrinked the css file from  3347 lines to 257 lines

IMPORTANT : Do not merge this in production. This is a milestone commit to denote completion of Desktop version.