[TABLET & PHONE VERSION] :: Fixed Design Issues
Fixed all the design issues with the website and aligned the front-end code to replicate the desgin patterns as per the PSD.

Rewrote the CSS from scratch to enfore DRY. Shrinked the css file from  3347 lines to 257 lines

This can be merged in production. This milestone commit denotes completion of Tablet & Phone version.